+++ docs/src/releases.md
@@ -4,6 +4,12 @@
 
 ### Version 1.0
 
+#### Version 1.0.2 (28 March, 2025)
+
+- Enable the normalisation of the observables. 
+- Add uniform spatial integral and component filter of the observables. 
+- Add conversion from Term to STerm.
+
 #### Version 1.0.1 (11 March, 2025)
 
 - Implement the pure bosonic models.